fix Issue #8
After setLocalTime() (GMThour == hours) is always true, but GMTtime might still be at another hour. If you return from here, the new GMTtime will not be set in line 307, so GMThour and GMTtime might have different hours. This produced the error in Issue #8 on 2023-10-01. The time was set to 23-00-00, but the above if statement was true and GMTtime not set as expected. The error did not occur with 23-00-01.